+++ src/Kalnoy/Cruddy/Schema/Fields/InlineTypes/HasOne.php
@@ -1,0 +1,26 @@
+<?php
+
+namespace Kalnoy\Cruddy\Schema\Fields\InlineTypes;
+
+use Illuminate\Database\Eloquent\Model;
+use Kalnoy\Cruddy\Schema\Fields\InlineRelation;
+
+/**
+ * This field will allow to inlinely edit related model.
+ *
+ * @since 1.0.0
+ */
+class HasOne extends InlineRelation {
+
+    /**
+     * @param Model $model
+     * @param Model $parent
+     */
+    public function joinModels(Model $model, Model $parent)
+    {
+        parent::joinModels($model, $parent);
+
+        $model->setAttribute($this->relation->getPlainForeignKey(), $parent->getKey());
+    }
+
+}